backend/requirements.txt
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.8

# Useful for handling environment variables
python-dotenv==1.0.0 

# Django CORS Headers for handling cross-origin requests in development
django-cors-headers==3.14.0

djangorestframework-simplejwt==5.2.2